backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.0
groq==0.20.0
pydantic==2.10.6
python-multipart==0.0.9
